implemented pundit for contents

+++ app/services/contents/delete_content_service.rb
@@ -2,15 +2,16 @@
 
 module Contents
   class DeleteContentService < BaseService
-    attr_reader :current_user, :content_params
+    attr_reader :current_user, :id
 
-    def initialize(current_user, content_params)
+    def initialize(current_user, id)
       @current_user = current_user
-      @content_params = content_params
+      @id = id
     end
 
     def call
-      content = current_user.contents.find_by!(content_params.extract!(:id))
+      content = Content.find(id)
+      authorize content, :destroy?
 
       content.destroy
